bot/requirements.txt
paramiko==3.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-telegram-bot==13.7
